chore(suite): recovery middleware
