--- packages/suite/src/middlewares/recovery/index.ts
@@ -1,3 +1,0 @@
-import recovery from './recoveryMiddleware';
-
-export default [recovery];
