+++ suite/recovery/src/index.ts
@@ -1,4 +1,5 @@
 export * from './isRecoveryInProgress';
+export * from './recoveryMiddleware';
 export * from './recoveryReducer';
 export * from './recoverySelectors';
 export * from './recoveryThunks';
